read_impl
